feat: refactor anvil server crate
