+++ anvil/anvil-server/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "anvil-server"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+description = """
+Customizable RPC server
+"""
+
+[dependencies]
+anvil-rpc = { path = "../anvil-rpc" }
+
+# axum related
+axum = { version = "0.5", features = ["ws"] }
+hyper = "0.14"
+tower-http = { version = "0.2", features = ["trace"] }
+
+# tracing
+tracing = "0.1.32"
+
+# async
+parking_lot = "0.12"
+futures = "0.3"
+
+# misc
+serde_json = "1.0.67"
+serde = { version = "1.0.136", features = ["derive"] }
+async-trait = "0.1.53"
